+++ src/Rougamo.Fody/Signature/Patterns/ITypePattern.cs
@@ -4,6 +4,8 @@
     {
         bool IsAny { get; }
 
+        bool IsVoid { get; }
+
         bool AssignableMatch { get; }
 
         bool IsMatch(TypeSignature signature);