+++ src/Rougamo.Fody/Signature/Patterns/NamespacePattern.cs
@@ -6,7 +6,6 @@ namespace Rougamo.Fody.Signature.Patterns
 {
     public class NamespacePattern : INamespacePattern
     {
-        private readonly string[] _patterns;
         private readonly int _ellipsisCount;
 
         public NamespacePattern(TokenSource tokens)
@@ -40,26 +39,28 @@ namespace Rougamo.Fody.Signature.Patterns
                 }
             }
             if (pattern.Length != 0) patterns.Add(pattern);
-            _patterns = patterns.ToArray();
+            Patterns = patterns.ToArray();
         }
+
+        public string[] Patterns { get; }
 
         public bool IsMatch(string @namespace)
         {
             var nss = @namespace.Split(new[] { '.' }, StringSplitOptions.RemoveEmptyEntries);
-            if (_ellipsisCount == 0 && nss.Length != _patterns.Length) return false;
-            if (_patterns.Length == 2 && _patterns[0] == "*" && _patterns[1] == Token.ELLIPSIS) return true;
+            if (_ellipsisCount == 0 && nss.Length != Patterns.Length) return false;
+            if (Patterns.Length == 2 && Patterns[0] == "*" && Patterns[1] == Token.ELLIPSIS) return true;
 
             int i = 0, j = 0;
             int iEllipsis = -1, jEllipsis = -1;
 
             while (i < nss.Length)
             {
-                if (j < _patterns.Length && WildcardMatcher.IsMatch(_patterns[j], nss[i]))
+                if (j < Patterns.Length && WildcardMatcher.IsMatch(Patterns[j], nss[i]))
                 {
                     i++;
                     j++;
                 }
-                else if (j < _patterns.Length && _patterns[j] == Token.ELLIPSIS)
+                else if (j < Patterns.Length && Patterns[j] == Token.ELLIPSIS)
                 {
                     iEllipsis = i;
                     jEllipsis = j + 1;
@@ -76,12 +77,12 @@ namespace Rougamo.Fody.Signature.Patterns
                 }
             }
 
-            while (j < _patterns.Length && _patterns[j] == Token.ELLIPSIS)
+            while (j < Patterns.Length && Patterns[j] == Token.ELLIPSIS)
             {
                 j++;
             }
 
-            return j == _patterns.Length;
+            return j == Patterns.Length;
         }
     }
 }