+++ test/TestAssemblies/SignatureUsage/Assignables/ImplBaseClass.cs
@@ -1,4 +1,6 @@
-﻿namespace SignatureUsage.Assignables
+﻿using System.Threading.Tasks;
+
+namespace SignatureUsage.Assignables
 {
     public class ImplBaseClass : BaseClass
     {
@@ -9,5 +11,13 @@
         public BaseClass GetBase() => default;
 
         public override Interface GetInterface() => default;
+
+        public Task<string> GetStringAsync() => Task.FromResult(string.Empty);
+
+        public async ValueTask<int> GetIntAsync()
+        {
+            await Task.Yield();
+            return 0;
+        }
     }
 }