+++ test/TestAssemblies/SignatureUsage/Assignables/ImplInterfaceBaseClass.cs
@@ -1,4 +1,6 @@
-﻿namespace SignatureUsage.Assignables
+﻿using System.Threading.Tasks;
+
+namespace SignatureUsage.Assignables
 {
     public class ImplInterfaceBaseClass : BaseClass, Interface
     {
@@ -7,5 +9,11 @@
         public override Interface GetInterface() => default;
 
         public AbstractClass GetAbstract() => default;
+
+        public Task CompletedAsync() => Task.CompletedTask;
+
+        private async Task YieldAsync() => await Task.Yield();
+
+        public ValueTask DefaultAsync() => new ValueTask();
     }
 }